Show Unity log viewer if fail to generate schema
